improved types
