+++ src/core/Button.tsx
@@ -4,7 +4,7 @@ import '@vkontakte/vkui/dist/vkui.css'
 import './static/modal.css'
 import './static/style.css'
 
-import { DeLabButtonConfig } from './types/index'
+import { DeLabButtonConfig } from './types/react'
 
 const white = 'https://ipfs.io/ipfs/bafkreigpmboyvo43fa4ybalflby3pb3eg2emgzn7axkgd7rmvrgdpx4oja'
 const black = 'https://ipfs.io/ipfs/bafkreibbn3nq6avodph3lcg6qlak6tbjha7levxzwgyk7nyrwot3ajvuwq'
